# [WIP] Wave read buffers as list of vectors