Hi Fangjun @csukuangfj,
i continued the work on the idea of loading wav files to `std::list<std::vector<char>>` buffers.

The `vector::insert` was replaced by memcpy, and the sizes of set-of-buffers 
now copy the series of powers of two, as : 10, 20, 40, 80, ... MBs.

The speedup on 600MB wav file was 5% (10s -> 9.5s with 6 reads).

By setting the kBlockSize to 1GB (load everything to 1 vector, and `swap()` 
it to the output vector, the speedup was ~15% (10s -> 8.55s).

This seems to be the upper bound for the possibility of optimization.
And similar effect would be achieved by filling the final `Matrix<float>` 
directly from the `std::list<std::vector<char>>` set-of-buffers 
(there is one less memory copy).

Would you be interested in having this optimization in your repo ?

The current version in `master` is simple. And with the optimization 
it would be 15% faster on very long wave files (~3 hours). The loading 
speed of "normal" files should stay about the same...

I am asking, before rewriting the code again.

Best regards,
Karel.