@vesis84 2023-08-29
@csukuangfj 

I am more inclined towards continuing to use a single reading method, as the detection of streaming data is not 100% reliable.

For this reason the JHU guys decided to read the inputs per 1MB blocks even for the seemingly non-streaming audio files. This is to avoid excessively big allocations if some new "crazy" constant for "size of data block with audio samples" appears in the "data in the wild" audio files.

And, perhaps the 10MB buffer is already too big (5.5 mins of audio), 1MB would be enough (33 seconds).

And yes, I'd like to remove the extra copy, this is the goal of the next iteration of rewriting...

K.