@vesis84 2023-08-29
The code is rewritten, now there is no 2nd copy of the audio buffer, and it is in `std::list<std::vector<char>>` buffers.
But for some reason the reading is slower. Maybe because of the more complicated for loop for exporting samples to Matrix.

For the moment I give up...

And the code is here:
https://github.com/vesis84/kaldi_native_io/tree/kaldi_native_io_PR_1COPY_SLOW